Move some items for organization purposes
