+++ app/Http/Controllers/Frontend/User/ProfileController.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Domains\Auth\Http\Controllers\Frontend\User;
+namespace App\Http\Controllers\Frontend\User;
 
-use App\Domains\Auth\Http\Requests\Frontend\User\UpdateProfileRequest;
+use App\Http\Requests\Frontend\User\UpdateProfileRequest;
 use App\Domains\Auth\Services\UserService;
 use App\Http\Controllers\Controller;
