+++ app/Http/Requests/Frontend/User/UpdateProfileRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Domains\Auth\Http\Requests\Frontend\User;
+namespace App\Http\Requests\Frontend\User;
 
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
