+++ routes/frontend/home.php
@@ -1,8 +1,8 @@
 <?php
 
-use App\Domains\Auth\Http\Controllers\Frontend\User\AccountController;
-use App\Domains\Auth\Http\Controllers\Frontend\User\DashboardController;
-use App\Domains\Auth\Http\Controllers\Frontend\User\ProfileController;
+use App\Http\Controllers\Frontend\User\AccountController;
+use App\Http\Controllers\Frontend\User\DashboardController;
+use App\Http\Controllers\Frontend\User\ProfileController;
 use App\Domains\Auth\Models\User;
 use App\Http\Controllers\Frontend\HomeController;
 use App\Http\Controllers\Frontend\TermsController;
